refactor: split cli.ts into focused modules
Extract domain/IP utilities to src/domain-utils.ts, API proxy
config logic to src/api-proxy-config.ts, and input parsers to
src/option-parsers.ts. The main cli.ts retains Commander setup
and action handlers only.

Closes #2472
